+++ demo/src/main/java/com/example/IDefensa.java
@@ -3,5 +3,6 @@ package com.example;
 // Interfaz para defensas (ejemplo: Escudo)
 public interface IDefensa {
     // Método que recibe un daño entrante y devuelve el daño neto tras aplicar la defensa
-    int aplicar(int danioEntrante);
+    // Usamos double para permitir daño fraccionado (ej: 0.5)
+    double aplicar(double danioEntrante);
 }
